backend/requirements.txt
baml-py==0.90.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.7.4
pyjwt==2.8.0
cryptography==41.0.7 